+++ lib/session_jar.dart
@@ -1,9 +1,4 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library session_jar;
 
 export 'src/session_jar.dart';
 export 'src/session.dart';
-
-// TODO: Export any libraries intended for clients of this package.