+++ lib/src/session.dart
@@ -1,16 +1,20 @@
 import 'dart:async';
 import 'dart:io';
-
 import '../session_jar.dart';
 
 ///containes all the data of one session
 class Session<T> {
   final String token;
-  final T body;
+  T body;
   final SessionJar sessionJar;
 
-  DateTime? expiresOn;
-  Timer? expirationTimer;
+  DateTime? _expiresOn;
+
+  DateTime? get expiresOn => _expiresOn;
+
+  //this field is used, the timer starts ticking on creation of this class
+  // ignore: unused_field
+  Timer? _expirationTimer;
 
   ///returns the session as a [Cookie]. based on the settings of its [CookieJar].
   Cookie get cookie {
@@ -25,9 +29,15 @@ class Session<T> {
       required this.body,
       required this.sessionJar,
       Duration? expiresIn}) {
+    //sets the expirationtime to the default value if needed
+    if (expiresIn == null && sessionJar.defaultExpirationTime != null) {
+      expiresIn = sessionJar.defaultExpirationTime;
+    }
+    //starts the expirationtimer
     if (expiresIn != null) {
-      expiresOn = DateTime.now().add(expiresIn);
-      expirationTimer = Timer(expiresIn, () => sessionJar.deleteSession(token));
+      _expiresOn = DateTime.now().add(expiresIn);
+      _expirationTimer =
+          Timer(expiresIn, () => sessionJar.deleteSession(token));
     }
   }
 }